[BE] feat: 이미지 업로드 분기처리, 상품 평균 평점 기능 구현 (#94)
* refactor: 리뷰작성시 별점 1점대로만 허용 수정

* feat: 리뷰 등록될 때마다 평균 평점 업데이트 구현

* feat: 이미지 업로드 안할시 분기처리 구현

* refactor: productId 삭제

+++ backend/src/main/java/com/funeat/review/presentation/dto/ReviewCreateRequest.java
@@ -4,13 +4,13 @@ import java.util.List;
 
 public class ReviewCreateRequest {
 
-    private final Double rating;
+    private final Long rating;
     private final List<Long> tagIds;
     private final String content;
     private final Boolean reBuy;
     private final Long memberId;
 
-    public ReviewCreateRequest(final Double rating, final List<Long> tagIds, final String content, final Boolean reBuy,
+    public ReviewCreateRequest(final Long rating, final List<Long> tagIds, final String content, final Boolean reBuy,
                                final Long memberId) {
         this.rating = rating;
         this.tagIds = tagIds;
@@ -19,7 +19,7 @@ public class ReviewCreateRequest {
         this.memberId = memberId;
     }
 
-    public Double getRating() {
+    public Long getRating() {
         return rating;
     }
